correct installation instructions
Add missing nltk import and install pytorch first, before whisperspeech.  Verified that speechbrain's installation of pytorch does not override the cuda-version installed prior to it like I thought it might.  Therefore, no need to install whisperspeech first (which installs speechbrain), uninstall pytorch, and then reinstall a cuda (or other) version of pytorch.

Also, added specific version of libraries to install in case there are breaking changes in any future releases.  Can be changed to newer version if need be in the future of course.