Update to expand user path
This is a minor update to fix the case where `~` is not recognized as the user home directory.  Referencing https://docs.python.org/3/library/os.path.html "Unlike a Unix shell, Python does not do any automatic path expansions. Functions such as expanduser() and expandvars() can be invoked explicitly when an application desires shell-like path expansion."  This fix ensures cross-platform compatability for storing and loading items in the hidden cache folder of the home directory.